% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clevr.R
\docType{package}
\name{clevr-package}
\alias{clevr}
\alias{clevr-package}
\title{clevr: Clustering and Link Prediction Evaluation in R}
\description{
Tools for evaluating link prediction and clustering algorithms 
    with respect to ground truth. Includes efficient implementations of 
    common performance measures such as pairwise precision/recall, 
    cluster homogeneity/completeness, variation of information, 
    Rand index etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cleanzr/clevr}
  \item Report bugs at \url{https://github.com/cleanzr/clevr/issues}
}

}
\author{
\strong{Maintainer}: Neil Marchant \email{ngmarchant@gmail.com}

Authors:
\itemize{
  \item Rebecca Steorts \email{beka@stat.duke.edu}
}

}
\keyword{internal}
