% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{clevr}
\alias{clevr}
\title{clevr}
\description{
Tools for evaluating link prediction and clustering algorithms with
respect to ground truth. Includes efficient implementations of common
performance measures such as pairwise precision/recall, cluster
homogeneity/completeness, variation of information, Rand index etc.
}
\author{
Neil Marchant
}
